requirements.txt
anthropic>=0.40.0
playwright>=1.44.0
feedparser
pyyaml>=6.0
